runner モジュールは削除して subcommand_legacy に統合
